update in db query
