+++ api/src/routes/search_router.py
@@ -8,9 +8,9 @@ from services.auth_service import search_users
 logger = logging.getLogger(__name__)
 
 db = firestore.client()
-router = APIRouter()
+router = APIRouter(prefix="/search", tags=["search"])
 
 
-@router.get("/search/search-users")
+@router.get("/search-users")
 def search(query: str = Query(..., min_length=1)):
     return search_users(query)
